requirements.txt
# For running the application
pywebview
Flask

# Optional: For SSL support
cryptography

# For building the application with PyInstaller
pyinstaller
rich
